+++ src/js/ApiServices.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import genresObj from '../db/genres.json';
 
 const API_KEY = 'api_key=afc2024e71269fd3f1e3cbd71f7c4df4';
 const BASE_URL = 'https://api.themoviedb.org/3/';
@@ -56,5 +57,24 @@ export default class ApiServices {
   }
   setPage(page) {
     this.page = page;
+  }
+
+  transformData(data) {
+    return data.map(film => {
+      film.year = film.release_date ? film.release_date.split('-')[0] : 'No information';
+
+      const genresNamesArr = film.genre_ids.map(
+        id => genresObj.genres.find(item => item.id === id).name,
+      );
+
+      if (genresNamesArr.length === 0) {
+        film.genres = 'No information';
+      } else if (genresNamesArr.length <= 3) {
+        film.genres = genresNamesArr.join(', ');
+      } else {
+        film.genres = `${genresNamesArr[0]}, ${genresNamesArr[1]}, Other`;
+      }
+      return film;
+    });
   }
 }
