--- src/js/film-card-transform-data.js
@@ -1,22 +1,0 @@
-import genresObj from '../db/genres.json';
-
-// rename, refactored to method of ApiServices class
-export function filmCardTransformData(data) {
-  // console.log(genresObj);
-  return data.map(film => {
-    film.year = film.release_date ? film.release_date.split('-')[0] : 'No information';
-
-    const genresNamesArr = film.genre_ids.map(
-      id => genresObj.genres.find(item => item.id === id).name,
-    );
-
-    if (genresNamesArr.length === 0) {
-      film.genres = 'No information';
-    } else if (genresNamesArr.length <= 3) {
-      film.genres = genresNamesArr.join(', ');
-    } else {
-      film.genres = `${genresNamesArr[0]}, ${genresNamesArr[1]}, Other`;
-    }
-    return film;
-  });
-}
